Refactor next 3 containers into components
- AvailableContainer
- DrawerContainer
- NavContainer
