+++ app/components/AvailableDashboard.jsx
@@ -1,7 +1,11 @@
 import React from 'react'
+import { connect } from 'react-redux'
+
 import AvailableIssueItem from './AvailableIssueItem'
 
-export default ({ issues }) => {
+/* ----------------- COMPONENT ------------------ */
+
+const AvailableDashboard = ({ issues }) => {
   return (
     <main className="mdl-layout__content dashboard">
       <div className="page-content mdl-cell mdl-cell--6-col">
@@ -26,3 +30,13 @@ export default ({ issues }) => {
     </main>
   )
 }
+
+/* ----------------- CONTAINER ------------------ */
+
+const mapStateToProps = state => {
+  return {
+    issues: state.issues
+  }
+}
+
+export default connect(mapStateToProps)(AvailableDashboard)
